requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-groq
chromadb
groq
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
streamlit
langchain-ollama